refactor: remove circular dependencies in `ui/pages/snap-account-redirect/*`
