+++ development/circular-deps.jsonc
@@ -244,13 +244,6 @@
     "ui/pages/notifications/notifications.tsx"
   ],
   [
-    "ui/pages/snap-account-redirect/components/index.ts",
-    "ui/pages/snap-account-redirect/components/snap-account-redirect-context.tsx",
-    "ui/pages/snap-account-redirect/components/snap-account-redirect-message.tsx",
-    "ui/pages/snap-account-redirect/components/url-display-box.tsx",
-    "ui/pages/snap-account-redirect/snap-account-redirect.tsx"
-  ],
-  [
     "ui/pages/swaps/swaps.util.ts",
     "ui/store/actions.ts"
   ]
